Compiles after moving types.proto to types module.
